Refactor PDF functions into separate file / module
