+++ src/docgen/docgen.js
@@ -2,21 +2,17 @@ import chalk from 'chalk';
 import path from 'path';
 import cheerio from 'cheerio';
 import moment from 'moment';
-import { spawn, exec } from 'child_process';
-import spawnArgs from 'spawn-args';
 import MarkdownIt from 'markdown-it';
 import imageSizeOf from 'image-size';
-import { Spinner as cliSpinner } from 'cli-spinner';
 import {
   readFile,
   writeFile,
   copyDirectory,
-  removeDirectory,
   cleanDirectory,
   makeDirectory,
 } from './fs/fs';
 import { validateJSON } from './validation/validation';
-import { execute } from './execute/execute';
+import { checkPdfVersion, generatePdf } from './pdf/wkhtmltopdf/wkhtmltopdf';
 import { version } from '../../package.json';
 
 const markdown = new MarkdownIt('commonmark').enable('table');
@@ -30,7 +26,6 @@ markdown.validateLink = () => {
 
 export function DocGen(process) {
   let mainProcess = process;
-  let wkhtmltopdfVersion = 'wkhtmltopdf 0.12.6 (with patched qt)'; //output from wkhtmltopdf -V
   let options;
   let templates = {};
   let meta = {};
@@ -81,8 +76,8 @@ export function DocGen(process) {
     await writePages();
     await createRedirect();
     if (options.pdf === true) {
-      await checkPdfVersion();
-      await generatePdf();
+      await checkPdfVersion({ options, mainProcess });
+      await generatePdf({ options, meta, sortedPages, mainProcess });
     }
     console.log(chalk.green.bold('Done!'));
   };
@@ -628,134 +623,6 @@ export function DocGen(process) {
       }
       mainProcess.exit(1);
     }
-  };
-
-  /*
-    wkthmltopdf options
-  */
-
-  let pdfOptions = [
-    ' --zoom 1.0',
-    ' --image-quality 100',
-    ' --print-media-type',
-    ' --orientation portrait',
-    ' --page-size A4',
-    ' --margin-top 25',
-    ' --margin-right 15',
-    ' --margin-bottom 16',
-    ' --margin-left 15',
-    ' --header-spacing 5',
-    ' --footer-spacing 5',
-    ' --no-stop-slow-scripts',
-  ];
-
-  let getPdfArguments = () => {
-    let pdfName = meta.parameters.name.toLowerCase() + '.pdf';
-    pdfOptions.push(' --enable-local-file-access');
-    pdfOptions.push(' --javascript-delay ' + options.pdfDelay); //code syntax highlight in wkhtmltopdf 0.12.2.1 fails without a delay (but why doesn't --no-stop-slow-scripts work?)
-    pdfOptions.push(
-      ' --user-style-sheet ' + __dirname + '/../include/pdf-stylesheet.css',
-    );
-    pdfOptions.push(' --header-html ' + options.output + 'temp/pdfHeader.html');
-    pdfOptions.push(' --footer-html ' + options.output + 'temp/pdfFooter.html');
-    pdfOptions.push(' cover ' + options.output + 'temp/pdfCover.html');
-    pdfOptions.push(
-      ' toc --xsl-style-sheet ' + __dirname + '/../include/pdf-contents.xsl',
-    );
-    let allPages = '';
-    for (let key in sortedPages) {
-      if (sortedPages.hasOwnProperty(key)) {
-        sortedPages[key].forEach((section) => {
-          section.pages.forEach((page) => {
-            let key = page.source;
-            let name = key.substr(0, page.source.lastIndexOf('.'));
-            let path = options.output + name + '.html';
-            allPages += ' ' + path;
-          });
-        });
-      }
-    }
-    let args = pdfOptions.join('');
-    args += allPages;
-    args += ' ' + options.output + pdfName;
-    return spawnArgs(args);
-  };
-
-  const checkPdfVersion = async () => {
-    try {
-      const stdout = await execute(options.wkhtmltopdfPath + ' -V');
-      //warn if the version of wkhtmltopdf is not an expected version
-      const actualWkhtmltopdfVersion = stdout.trim();
-      if (actualWkhtmltopdfVersion !== wkhtmltopdfVersion) {
-        const warning =
-          'Warning: unexpected version of wkhtmltopdf, which may work but is not tested or supported';
-        const expectedVersion = '   expected version: ' + wkhtmltopdfVersion;
-        const detectedVersion =
-          '   detected version: ' + actualWkhtmltopdfVersion;
-        console.log(chalk.yellow(warning));
-        console.log(chalk.yellow(expectedVersion));
-        console.log(chalk.yellow(detectedVersion));
-      }
-    } catch (error) {
-      console.log(
-        chalk.red(
-          'Unable to call wkhtmltopdf. Is it installed and in path? See http://wkhtmltopdf.org',
-        ),
-      );
-      if (options.verbose === true) {
-        console.log(chalk.red(error));
-      }
-      mainProcess.exit(1);
-    }
-  };
-
-  /*
-    call wkhtmltopdf as an external executable
-  */
-
-  let generatePdf = async () => {
-    console.log(chalk.green('Creating the PDF copy (may take some time)'));
-    let args = getPdfArguments();
-    let wkhtmltopdf = spawn(options.wkhtmltopdfPath, args);
-    let spinner = new cliSpinner(chalk.green('   Processing... %s'));
-    spinner.setSpinnerString('|/-\\');
-
-    wkhtmltopdf.on('error', (error) => {
-      console.log(chalk.red('Error calling wkhtmltopdf to generate the PDF'));
-      if (options.verbose === true) {
-        console.log(chalk.red(error));
-      }
-    });
-
-    if (options.verbose !== true) {
-      spinner.start(); //only show spinner when verbose is off (otherwise show raw wkhtmltopdf output)
-    } else {
-      //pipe the output from wkhtmltopdf back to stdout
-      //however, wkhtmltpdf outputs to stderr, not stdout. See:
-      //https://github.com/wkhtmltopdf/wkhtmltopdf/issues/1980
-      wkhtmltopdf.stderr.pipe(mainProcess.stdout);
-    }
-
-    wkhtmltopdf.stdout.on('data', (data) => {
-      //do nothing
-    });
-
-    wkhtmltopdf.stderr.on('data', (data) => {
-      //do nothing
-    });
-
-    wkhtmltopdf.on('close', async (code) => {
-      if (options.verbose !== true) {
-        spinner.stop();
-        console.log(''); //newline after spinner stops
-      }
-      if (code !== 0) {
-        let warning =
-          'wkhtmltopdf exited with a warning or error: try the -v option for details';
-        console.log(chalk.yellow(warning));
-      }
-      await removeDirectory(options.output + 'temp');
-    });
   };
 
   let createRedirect = async () => {
